refactor(core): create consolidate_block function
